Move the relative path conversion function for common use
